Dropdown and Recent Language Section
This commit implements the top parts of the text area and display area section. The top part includes 3 recently used languages along with a dropdown arrow which displays all the supported languages that the users can choose from. This dropdown section also includes a search functionality so that the users can easily search up the language they want to translate in.